# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_daidat02_server_internal_domain_product.CreateProductRequest:
    properties:
      category_id:
        type: string
      cost_price:
        type: number
      is_active:
        type: boolean
      min_stock:
        type: number
      name:
        type: string
      net_amount:
        type: number
      net_unit:
        type: string
      product_type:
        type: string
      sell_price:
        type: number
      show_on_web:
        type: boolean
      sku:
        type: string
      stock_quantity:
        type: number
      unit:
        type: string
    type: object
  github_com_daidat02_server_internal_domain_product.UpdateProductRequest:
    properties:
      category_id:
        type: string
      cost_price:
        type: number
      is_active:
        type: boolean
      min_stock:
        type: number
      name:
        type: string
      net_amount:
        type: number
      net_unit:
        type: string
      product_type:
        type: string
      sell_price:
        type: number
      show_on_web:
        type: boolean
      sku:
        type: string
      stock_quantity:
        type: number
      unit:
        type: string
    type: object
...
  github_com_daidat02_server_internal_domain_staff.CreateStaffRequest:
    properties:
      code:
        type: string
      commission_rate:
        type: number
      full_name:
        type: string
      hire_date:
        type: string
      id:
        type: string
      organization_id:
        type: string
      phone:
        type: string
      position:
        type: string
      status:
        type: string
    type: object
  github_com_daidat02_server_internal_domain_staff.UpdateStaffRequest:
    properties:
      code:
        type: string
      commission_rate:
        type: number
      full_name:
        type: string
      hire_date:
        type: string
      phone:
        type: string
      position:
        type: string
      status:
        type: string
    type: object
...
  github_com_daidat02_server_pkg_response.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  github_com_daidat02_server_pkg_response.SuccessResponse:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 'API quản lý salon: user, nhân viên, khách hàng.'
  title: Salon Management API
  version: "1.0"
paths:
...
  /products:
    get:
      parameters:
      - description: Trang (mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số dòng/trang (mặc định 20, tối đa 100)
        in: query
        name: page_size
        type: integer
      - description: Tìm theo tên, SKU hoặc loại sản phẩm (ILIKE)
        in: query
        name: search
        type: string
      - description: Lọc trạng thái
        enum:
        - active
        - low_stock
        - out_of_stock
        in: query
        name: status
        type: string
      - description: Sắp xếp
        enum:
        - created_at_desc
        - price_asc
        - price_desc
        - stock_asc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách sản phẩm kèm phân trang {data, page, page_size, total,
            total_pages}
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "500":
          description: Lỗi server
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách sản phẩm (phân trang + lọc + sắp xếp)
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Thông tin sản phẩm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_daidat02_server_internal_domain_product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo sản phẩm thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Dữ liệu không hợp lệ / thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "500":
          description: Lỗi server (vd trùng SKU)
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo sản phẩm
      tags:
      - Products
  /products/{id}:
    delete:
      parameters:
      - description: ID sản phẩm
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa sản phẩm thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu id
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "404":
          description: Không tìm thấy sản phẩm
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa sản phẩm
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: ID sản phẩm
        in: path
        name: id
        required: true
        type: string
      - description: Thông tin cập nhật
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_daidat02_server_internal_domain_product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cập nhật sản phẩm thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu id / dữ liệu không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "404":
          description: Không tìm thấy sản phẩm
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật sản phẩm
      tags:
      - Products
...
  /staff:
    get:
      parameters:
      - description: Trang (mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số dòng/trang (mặc định 20, tối đa 100)
        in: query
        name: page_size
        type: integer
      - description: Tìm theo tên, mã, SĐT hoặc vị trí (ILIKE)
        in: query
        name: search
        type: string
      - description: Lọc trạng thái
        enum:
        - active
        - inactive
        in: query
        name: status
        type: string
      - description: Sắp xếp
        enum:
        - created_at_desc
        - name_asc
        - name_desc
        - hire_date_asc
        - hire_date_desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách nhân viên kèm phân trang {data, page, page_size,
            total, total_pages}
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "500":
          description: Lỗi server
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách nhân viên (phân trang + lọc + sắp xếp)
      tags:
      - Staff
    post:
      consumes:
      - application/json
      parameters:
      - description: Thông tin nhân viên
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_daidat02_server_internal_domain_staff.CreateStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo nhân viên thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Dữ liệu không hợp lệ / thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "500":
          description: Lỗi server
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo nhân viên
      tags:
      - Staff
  /staff/{id}:
    delete:
      parameters:
      - description: ID nhân viên
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa nhân viên thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu id
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "404":
          description: Không tìm thấy nhân viên
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa mềm nhân viên
      tags:
      - Staff
    put:
      consumes:
      - application/json
      parameters:
      - description: ID nhân viên
        in: path
        name: id
        required: true
        type: string
      - description: Thông tin cập nhật
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_daidat02_server_internal_domain_staff.UpdateStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cập nhật nhân viên thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu id / dữ liệu không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "404":
          description: Không tìm thấy nhân viên
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật nhân viên
      tags:
      - Staff
securityDefinitions:
  BearerAuth:
    description: '"Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"